requirements-NM.txt
aioify
instantmusic
telethon-aio
youtube-dl
eyed3
